lib/ukschedcoop: Implement ISR variant of `thread_woken_func`
This commit compiles the implementation of `schedcoop_thread_woken()`
as ISR-safe. The same function is used for both: non-ISR and ISR
contexts. This is safe to do because by definition an ISR-safe function
can be called from non-ISR context but not vice versa.
