+++ lib/ukschedcoop/Makefile.uk
@@ -4,3 +4,4 @@ CINCLUDES-$(CONFIG_LIBUKSCHEDCOOP)     += -I$(LIBUKSCHEDCOOP_BASE)/include
 CXXINCLUDES-$(CONFIG_LIBUKSCHEDCOOP)   += -I$(LIBUKSCHEDCOOP_BASE)/include
 
 LIBUKSCHEDCOOP_SRCS-y += $(LIBUKSCHEDCOOP_BASE)/schedcoop.c
+LIBUKSCHEDCOOP_SRCS-y += $(LIBUKSCHEDCOOP_BASE)/isrwoken.c|isr
